+++ packages/button/src/CloseButton.ts
@@ -57,7 +57,6 @@ const crossIcon: Record<string, () => TemplateResult> = {
  * @element sp-close-button
  *
  * @slot - text label of the Close Button
- * @slot icon - The icon to use for Close Button
  */
 export class CloseButton extends SizedMixin(StyledButton, {
     noDefaultSize: true,