+++ packages/menu/test/menu.test.ts
@@ -16,6 +16,7 @@ import {
     expect,
     html,
     nextFrame,
+    oneEvent,
     waitUntil,
 } from '@open-wc/testing';
 import { Menu, MenuItem } from '@spectrum-web-components/menu';
@@ -543,7 +544,9 @@ describe('Menu', () => {
         expect(secondItem.getAttribute('aria-checked')).to.equal('false');
         expect(el.value).to.equal('First');
 
+        const change = oneEvent(el, 'change');
         secondItem.click();
+        await change;
 
         await elementUpdated(el);
         await elementUpdated(firstItem);
@@ -655,7 +658,9 @@ describe('Menu', () => {
         expect(el.value).to.equal('First');
         expect(el.selectedItems.length).to.equal(1);
 
+        let change = oneEvent(el, 'change');
         secondItem.click();
+        await change;
 
         await elementUpdated(el);
         await elementUpdated(firstItem);
@@ -669,7 +674,9 @@ describe('Menu', () => {
         expect(el.value).to.equal('First,Second');
         expect(el.selectedItems.length).to.equal(2);
 
+        change = oneEvent(el, 'change');
         firstItem.click();
+        await change;
 
         await elementUpdated(el);
         await elementUpdated(firstItem);