+++ packages/number-field/src/NumberField.ts
@@ -34,13 +34,9 @@ import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron200.js';
 import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron50.js';
 import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron75.js';
 import '@spectrum-web-components/infield-button/sp-infield-button.js';
-import {
-    isAndroid,
-    isIOS,
-    isIPhone,
-} from '@spectrum-web-components/shared/src/platform.js';
 import { TextfieldBase } from '@spectrum-web-components/textfield';
 import styles from './number-field.css.js';
+import { isAndroid, isIOS, isIPhone } from '@spectrum-web-components/shared';
 
 export const FRAMES_PER_CHANGE = 5;
 // Debounce duration for inserting a `change` event after a batch of `wheel` originating `input` events.
@@ -376,11 +372,15 @@ export class NumberField extends TextfieldBase {
         this._value = this.validateInput(value);
         this.inputElement.value = this.numberFormatter.format(value);
 
-        this.inputElement.dispatchEvent(
-            new Event('input', { bubbles: true, composed: true })
-        );
+        const inputEvent = new Event('input', {
+            bubbles: true,
+            composed: true,
+        });
+        this.inputElement.readOnly = true;
+        this.inputElement.dispatchEvent(inputEvent);
         this.indeterminate = false;
         this.focus();
+        this.inputElement.readOnly = false;
     }
 
     private increment(factor = 1): void {