+++ packages/picker/src/Picker.ts
@@ -320,9 +320,6 @@ export class PickerBase extends SizedMixin(SpectrumElement, {
     ): Promise<void> {
         this.open = false;
         // should always close when "setting" a value
-        if (this.strategy) {
-            this.strategy.open = false;
-        }
         const oldSelectedItem = this.selectedItem;
         const oldValue = this.value;
 
@@ -503,13 +500,6 @@ export class PickerBase extends SizedMixin(SpectrumElement, {
                         this.size as DefaultElementSize
                     ]}"
                 ></sp-icon-chevron100>
-                <slot
-                    aria-hidden="true"
-                    name="tooltip"
-                    id="tooltip"
-                    @keydown=${this.handleKeydown}
-                    @slotchange=${this.handleTooltipSlotchange}
-                ></slot>
             `,
         ];
     }
@@ -604,6 +594,13 @@ export class PickerBase extends SizedMixin(SpectrumElement, {
             >
                 ${this.buttonContent}
             </button>
+            <slot
+                aria-hidden="true"
+                name="tooltip"
+                id="tooltip"
+                @keydown=${this.handleKeydown}
+                @slotchange=${this.handleTooltipSlotchange}
+            ></slot>
             ${this.renderMenu} ${this.renderDescriptionSlot}
         `;
     }
@@ -902,6 +899,16 @@ export class PickerBase extends SizedMixin(SpectrumElement, {
 
     public override connectedCallback(): void {
         super.connectedCallback();
+        this.updateComplete.then(() => {
+            if (!this.tooltipEl?.selfManaged) {
+                return;
+            }
+            const overlayElement = this.tooltipEl.overlayElement;
+            if (overlayElement) {
+                overlayElement.triggerElement = this.button;
+            }
+        });
+
         this.recentlyConnected = this.hasUpdated;
         this.addEventListener('focus', this.handleFocus);
     }